Handle Controller & Trials
Finished the handle controller, and started on the functionality for trials. Made a couple changes to the pinch controller, mainly putting in the first tap to start the selection. The biggest problem with trials right now is that the program crashes when a trial's goal is met, throwing an unhelpful exception with no line number.